4to Semestre/PYTHON/laboratorio_usuarios/requirements.txt
psycopg[binary,pool]>=3.1
python-dotenv==1.0.1
